+++ application/modules/portal/controllers/ApiController.php
@@ -52,10 +52,18 @@ class ApiController extends Zend_Controller_Action
             ->setBody(Zend_Json_Encoder::encode($response));
     }
 
-
-    public function ccsdMetricsAction() {
+    public function ccsdMetricsAction()
+    {
         $this->_helper->layout->disableLayout();
         $this->_helper->viewRenderer->setNoRender();
+
+        // Count of published documents
+        try {
+            $nbOfPublications = Episciences_PapersManager::getPublishedPapersCount();
+        } catch (Zend_Db_Statement_Exception $e) {
+            $nbOfPublications = 0;
+        }
+
 
         // Count of Users for the current year, at the time of request
         try {
@@ -67,7 +75,7 @@ class ApiController extends Zend_Controller_Action
         header('Content-Type: application/json');
 
         try {
-            echo json_encode(['nbOfEpisciencesUsers' => $nbOfUsers], JSON_THROW_ON_ERROR);
+            echo json_encode(['nbOfEpisciencesUsers' => $nbOfUsers, 'nbOfEpisciencesPublications' => $nbOfPublications], JSON_THROW_ON_ERROR);
         } catch (JsonException $e) {
             $timeOut = 1800;
             header('HTTP/1.0 503 Service Unavailable');
@@ -75,8 +83,6 @@ class ApiController extends Zend_Controller_Action
             printf("Sorry, this service is not available at the moment. We must be busy fixing a problem, please try again in %s seconds", $timeOut);
         }
     }
-
-
 
     /**
      * Metrics for OpenAIRE formatted as Openmetrics